requirements.txt
pytest==8.3.3
pytest-playwright==0.5.0
playwright==1.48.0
python-dotenv==1.0.1
PyYAML==6.0.2
openai==1.51.2
numpy==2.1.1
pydantic==2.9.2
allure-pytest==2.13.5


